+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_test.dart
@@ -200,6 +200,222 @@ void main() {
       () => testDefines(customDefineWithValue: value),
     );
   }
+
+  test('CBuilder flags', () async {
+    await inTempDir(
+      // https://github.com/dart-lang/sdk/issues/40159
+      keepTemp: Platform.isWindows,
+      (tempUri) async {
+        final addCUri =
+            packageUri.resolve('test/cbuilder/testfiles/add/src/add.c');
+        const name = 'add';
+
+        final logMessages = <String>[];
+        final logger = createCapturingLogger(logMessages);
+
+        final buildConfig = BuildConfig(
+          outDir: tempUri,
+          packageRoot: tempUri,
+          targetArchitecture: Architecture.current,
+          targetOs: OS.current,
+          buildMode: BuildMode.release,
+          linkModePreference: LinkModePreference.dynamic,
+          cCompiler: CCompilerConfig(
+            cc: cc,
+            envScript: envScript,
+            envScriptArgs: envScriptArgs,
+          ),
+        );
+        final buildOutput = BuildOutput();
+
+        final flag = switch (buildConfig.targetOs) {
+          OS.windows => '/O2',
+          _ => '-O2',
+        };
+
+        final cbuilder = CBuilder.library(
+          sources: [addCUri.toFilePath()],
+          name: name,
+          assetId: name,
+          flags: [flag],
+        );
+        await cbuilder.run(
+          buildConfig: buildConfig,
+          buildOutput: buildOutput,
+          logger: logger,
+        );
+
+        final dylibUri = tempUri.resolve(Target.current.os.dylibFileName(name));
+
+        final dylib = DynamicLibrary.open(dylibUri.toFilePath());
+        final add = dylib.lookupFunction<Int32 Function(Int32, Int32),
+            int Function(int, int)>('add');
+        expect(add(1, 2), 3);
+
+        final compilerInvocation = logMessages.singleWhere(
+          (message) => message.contains(addCUri.toFilePath()),
+        );
+        expect(compilerInvocation, contains(flag));
+      },
+    );
+  });
+
+  test('CBuilder includes', () async {
+    await inTempDir(
+      // https://github.com/dart-lang/sdk/issues/40159
+      keepTemp: Platform.isWindows,
+      (tempUri) async {
+        final includeDirectoryUri =
+            packageUri.resolve('test/cbuilder/testfiles/includes/include');
+        final includesHUri = packageUri
+            .resolve('test/cbuilder/testfiles/includes/include/includes.h');
+        final includesCUri = packageUri
+            .resolve('test/cbuilder/testfiles/includes/src/includes.c');
+        const name = 'includes';
+
+        final buildConfig = BuildConfig(
+          outDir: tempUri,
+          packageRoot: tempUri,
+          targetArchitecture: Architecture.current,
+          targetOs: OS.current,
+          buildMode: BuildMode.release,
+          linkModePreference: LinkModePreference.dynamic,
+          cCompiler: CCompilerConfig(
+            cc: cc,
+            envScript: envScript,
+            envScriptArgs: envScriptArgs,
+          ),
+        );
+        final buildOutput = BuildOutput();
+
+        final cbuilder = CBuilder.library(
+          name: name,
+          assetId: name,
+          includes: [includeDirectoryUri.toFilePath()],
+          sources: [includesCUri.toFilePath()],
+        );
+        await cbuilder.run(
+          buildConfig: buildConfig,
+          buildOutput: buildOutput,
+          logger: logger,
+        );
+
+        expect(buildOutput.dependencies.dependencies, contains(includesHUri));
+
+        final dylibUri = tempUri.resolve(Target.current.os.dylibFileName(name));
+        final dylib = DynamicLibrary.open(dylibUri.toFilePath());
+        final x = dylib.lookup<Int>('x');
+        expect(x.value, 42);
+      },
+    );
+  });
+
+  test('CBuilder std', () async {
+    await inTempDir(
+      // https://github.com/dart-lang/sdk/issues/40159
+      keepTemp: Platform.isWindows,
+      (tempUri) async {
+        final addCUri =
+            packageUri.resolve('test/cbuilder/testfiles/add/src/add.c');
+        const name = 'add';
+        const std = 'c99';
+
+        final logMessages = <String>[];
+        final logger = createCapturingLogger(logMessages);
+
+        final buildConfig = BuildConfig(
+          outDir: tempUri,
+          packageRoot: tempUri,
+          targetArchitecture: Architecture.current,
+          targetOs: OS.current,
+          buildMode: BuildMode.release,
+          linkModePreference: LinkModePreference.dynamic,
+          cCompiler: CCompilerConfig(
+            cc: cc,
+            envScript: envScript,
+            envScriptArgs: envScriptArgs,
+          ),
+        );
+        final buildOutput = BuildOutput();
+
+        final stdFlag = switch (buildConfig.targetOs) {
+          OS.windows => '/std:$std',
+          _ => '-std=$std',
+        };
+
+        final cbuilder = CBuilder.library(
+          sources: [addCUri.toFilePath()],
+          name: name,
+          assetId: name,
+          std: std,
+        );
+        await cbuilder.run(
+          buildConfig: buildConfig,
+          buildOutput: buildOutput,
+          logger: logger,
+        );
+
+        final dylibUri = tempUri.resolve(Target.current.os.dylibFileName(name));
+
+        final dylib = DynamicLibrary.open(dylibUri.toFilePath());
+        final add = dylib.lookupFunction<Int32 Function(Int32, Int32),
+            int Function(int, int)>('add');
+        expect(add(1, 2), 3);
+
+        final compilerInvocation = logMessages.singleWhere(
+          (message) => message.contains(addCUri.toFilePath()),
+        );
+        expect(compilerInvocation, contains(stdFlag));
+      },
+    );
+  });
+
+  test('CBuilder cpp', () async {
+    await inTempDir((tempUri) async {
+      final helloWorldCppUri = packageUri.resolve(
+          'test/cbuilder/testfiles/hello_world_cpp/src/hello_world_cpp.cc');
+      if (!await File.fromUri(helloWorldCppUri).exists()) {
+        throw Exception('Run the test from the root directory.');
+      }
+      const name = 'hello_world_cpp';
+
+      final buildConfig = BuildConfig(
+        buildMode: BuildMode.release,
+        outDir: tempUri,
+        packageRoot: tempUri,
+        targetArchitecture: Architecture.current,
+        targetOs: OS.current,
+        // Ignored by executables.
+        linkModePreference: LinkModePreference.dynamic,
+        cCompiler: CCompilerConfig(
+          cc: cc,
+          envScript: envScript,
+          envScriptArgs: envScriptArgs,
+        ),
+      );
+      final buildOutput = BuildOutput();
+      final cbuilder = CBuilder.executable(
+        name: name,
+        sources: [helloWorldCppUri.toFilePath()],
+        cpp: true,
+      );
+      await cbuilder.run(
+        buildConfig: buildConfig,
+        buildOutput: buildOutput,
+        logger: logger,
+      );
+
+      final executableUri =
+          tempUri.resolve(Target.current.os.executableFileName(name));
+      expect(await File.fromUri(executableUri).exists(), true);
+      final result = await runProcess(
+        executable: executableUri,
+        logger: logger,
+      );
+      expect(result.exitCode, 0);
+      expect(result.stdout.trim(), endsWith('Hello world.'));
+    });
+  });
 }
 
 Future<void> testDefines({